re-add "parse_url" function
