+++ src/remap/function.rs
@@ -38,6 +38,7 @@ mod md5;
 mod now;
 mod only_fields;
 mod parse_duration;
+mod parse_url;
 mod sha1;
 mod sha2;
 mod sha3;
@@ -66,6 +67,7 @@ pub use format_timestamp::FormatTimestamp;
 pub use now::Now;
 pub use only_fields::OnlyFields;
 pub use parse_duration::ParseDuration;
+pub use parse_url::ParseUrl;
 pub use slice::Slice;
 pub use split::Split;
 pub use starts_with::StartsWith;
